c++/tests: move to a single test dir and test binary
The AFQMC tests are now in a single binary called test_afqmc.
This simplifies the build system and has the potential for some compile
time savings later on.

We use the Catch2 CMake integration to automatically add the individual
tests to CMake so running the test target works the same as before.

Use test_afqmc --list-test to find the individual tests and run the musing filters.
