--- src/AFQMC/Estimators/tests/CMakeLists.txt
@@ -1,26 +1,0 @@
-
-set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_UNIT_TEST_DIR})
-
-set(TEST_ID afqmc_estimators)
-add_executable( test_${TEST_ID} test_estimators.cpp)
-target_link_libraries(test_${TEST_ID} PRIVATE catch_main)
-set(UTEST_NAME unit_test_${TEST_ID})
-add_unit_test(${UTEST_NAME} "${PROJECT_UNIT_TEST_DIR}/test_${TEST_ID}")
-set_tests_properties(${UTEST_NAME} PROPERTIES WORKING_DIRECTORY ${PROJECT_UNIT_TEST_DIR})
-set_property(TEST ${UTEST_NAME} APPEND PROPERTY LABELS "afqmc")
-
-set(TEST_ID afqmc_observables_reference)
-add_executable(test_${TEST_ID} test_observables_reference.cpp)
-target_link_libraries(test_${TEST_ID} PRIVATE catch_main)
-set(UTEST_NAME unit_test_${TEST_ID})
-add_unit_test(${UTEST_NAME} "${PROJECT_UNIT_TEST_DIR}/test_${TEST_ID}")
-set_tests_properties(${UTEST_NAME} PROPERTIES WORKING_DIRECTORY ${PROJECT_UNIT_TEST_DIR})
-set_property(TEST ${UTEST_NAME} APPEND PROPERTY LABELS "afqmc")
-
-set(TEST_ID afqmc_estimator_handler)
-add_executable( test_${TEST_ID} test_estimator_handler.cpp)
-target_link_libraries(test_${TEST_ID} PRIVATE catch_main)
-set(UTEST_NAME unit_test_${TEST_ID})
-add_unit_test(${UTEST_NAME} "${PROJECT_UNIT_TEST_DIR}/test_${TEST_ID}")
-set_tests_properties(${UTEST_NAME} PROPERTIES WORKING_DIRECTORY ${PROJECT_UNIT_TEST_DIR})
-set_property(TEST ${UTEST_NAME} APPEND PROPERTY LABELS "afqmc")
