--- src/AFQMC/Hamiltonians/tests/CMakeLists.txt
@@ -1,18 +1,0 @@
-
-set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_UNIT_TEST_DIR})
-
-set(TEST_ID afqmc_hamiltonians)
-
-set( TEST_SRCS 
-	test_hamiltonian_factory.cpp
-        )
-
-add_executable(test_${TEST_ID} ${TEST_SRCS} )
-target_link_libraries(test_${TEST_ID} PRIVATE catch_main)
-
-# MAM: Enable THC tests on both CPU and GPU
-
-set(UTEST_NAME unit_test_${TEST_ID})
-add_unit_test(${UTEST_NAME} "${PROJECT_UNIT_TEST_DIR}/test_${TEST_ID}")
-set_tests_properties(${UTEST_NAME} PROPERTIES WORKING_DIRECTORY ${PROJECT_UNIT_TEST_DIR})
-set_property(TEST ${UTEST_NAME} APPEND PROPERTY LABELS "afqmc")
