+++ src/numerics/sandbox/contract_tests.cpp
@@ -32,7 +32,7 @@
 #include "nda/nda.hpp"
 #include "nda/blas.hpp"
 #include "nda/tensor.hpp"
-#include "utilities/test_common.hpp"
+#include "test_common.hpp"
 #include "numerics/device_kernels/kernels.h"
 #include "numerics/nda_functions.hpp"
 #include "utilities/Timer.hpp"
