+++ src/numerics/sandbox/mpi_tests.cpp
@@ -33,7 +33,7 @@
 #include "nda/nda.hpp"
 #include "numerics/distributed_array/nda.hpp"
 #include "numerics/shared_array/nda.hpp"
-#include "utilities/test_common.hpp"
+#include "test_common.hpp"
 #include "utilities/Timer.hpp"
 
 namespace bdft_tests
