+++ src/numerics/shared_array/tests/test_shared.cpp
@@ -31,7 +31,7 @@
 
 #include "nda/nda.hpp"
 #include "numerics/shared_array/nda.hpp"
-#include "utilities/test_common.hpp"
+#include "test_common.hpp"
 
 namespace bdft_tests
 {
