+++ src/numerics/tests/test_nda_functions.cpp
@@ -31,7 +31,7 @@
 
 #include "nda/nda.hpp"
 #include "nda/tensor.hpp"
-#include "utilities/test_common.hpp"
+#include "test_common.hpp"
 #include "numerics/device_kernels/kernels.h"
 #include "numerics/nda_functions.hpp"
 #include "utilities/Timer.hpp"
