+++ src/utilities/CMakeLists.txt
@@ -35,11 +35,3 @@ target_sources(safire_lib
     freemem.cpp
 )
 
-if(BUILD_UNIT_TESTS)
-  add_library(catch_main STATIC catch_main.cpp)
-  target_link_libraries(catch_main PUBLIC safire_lib Catch2::Catch2)
-  if(ENABLE_CPPTRACE)
-    target_link_libraries(safire_lib PUBLIC cpptrace::cpptrace)
-  endif()
-endif(BUILD_UNIT_TESTS)
-
