+++ tests/hubbard_factorizations.hpp
