+++ tests/test_estimators.cpp
@@ -23,7 +23,7 @@
 #include "IO/ptree/ptree_utilities.hpp"
 #include "utilities/Random.hpp"
 #include "utilities/check.hpp"
-#include "utilities/test_common.hpp"
+#include "test_common.hpp"
 #include "utilities/h5_utils.hpp"
 #include "IO/app_loggers.h"
